# Migrate rating system from Topcoder to Elo-MMR