--- a/judge/migrations/0123_contest_rating_elo_mmr.py
+++ b/judge/migrations/0123_contest_rating_elo_mmr.py
@@ -1,10 +1,10 @@
 import math
+from operator import attrgetter, itemgetter
 
 from django.db import migrations, models
 from django.db.models import Count, OuterRef, Subquery
 from django.db.models.functions import Coalesce
 from django.utils import timezone
-from operator import attrgetter, itemgetter
 
 
 def tie_ranker(iterable, key=attrgetter('points')):