--- a/judge/ratings.py
+++ b/judge/ratings.py
@@ -162,8 +162,7 @@ def rate_contest(contest):
 
 
 RATING_LEVELS = ['Newbie', 'Amateur', 'Expert', 'Candidate Master', 'Master', 'Grandmaster', 'Target']
-# TODO: adjust RATING_VALUES later
-RATING_VALUES = [1000, 1200, 1500, 1800, 2200, 3000]
+RATING_VALUES = [1100, 1400, 1700, 2000, 2400, 3000]
 RATING_CLASS = ['rate-newbie', 'rate-amateur', 'rate-expert', 'rate-candidate-master',
                 'rate-master', 'rate-grandmaster', 'rate-target']