--- a/judge/ratings.py
+++ b/judge/ratings.py
@@ -23,17 +23,24 @@ def eval_tanhs(tanh_terms, x):
 
 
 def solve(tanh_terms, y_tg, lin_factor=0, bounds=VALID_RANGE):
-    L,R = bounds
-    while R-L > 1e-5*SD_INIT:
+    L, R = bounds
+    Ly, Ry = None, None
+    while R-L > 2:
         x = (L+R) / 2
         y = lin_factor * x + eval_tanhs(tanh_terms, x)
         if y > y_tg:
-            R = x
+            R, Ry = x, y
         elif y < y_tg:
-            L = x
+            L, Ly = x, y
         else:
             return x
-    return (L+R) / 2
+    if Ly is None:
+        Ly = lin_factor * L + eval_tanhs(tanh_terms, L)
+    if Ry is None:
+        Ry = lin_factor * R + eval_tanhs(tanh_terms, R)
+    # Use linear interpolation to be slightly more accurate.
+    ratio = (y_tg-Ly) / (Ry-Ly) if Ly < y_tg < Ry else 0.5
+    return L*(1-ratio) + R*ratio
 
 
 def tie_ranker(iterable, key=attrgetter('points')):