[Elo-MMR](https://arxiv.org/abs/2101.00400) is a rating system that has more desirable properties than Topcoder's rating system. For example, Elo-MMR converges faster, incentivizes users to not [farm volatility](https://codeforces.com/blog/entry/87848), and has a more reasonable performance calculation.

Change in the rating curve (ratings rounded to the nearest 100x+50): ![ratingcurve](https://user-images.githubusercontent.com/14223529/129460372-50a46bc4-acb0-4140-8b16-73b6d0ef2d35.PNG)

Change in each user's rating (i.e. users can expect their rating to increase by +100): 
![ratingdelta](https://user-images.githubusercontent.com/14223529/129460462-0f44a46f-34d7-4a03-9209-7c8a1c05a385.PNG)

Due to the expected +100 rating increase, I modified the rating cutoffs:
- Old: `RATING_VALUES = [1000, 1200, 1500, 1800, 2200, 3000]`
- New: `RATING_VALUES = [1000, 1300, 1600, 1900, 2400, 3000]`

Change in performance (on my virtual machine):
- It takes 8 seconds to rate all 111 contests using Topcoder.
- It takes 14 seconds to rate all 111 contests using Elo-MMR. The biggest bottlenecks are `solve()` (5 seconds), and the query for `historical_p` (3 seconds).